requirements.txt
appnope==0.1.4
asttokens==2.4.1
backports.tarfile==1.2.0
black==24.4.2
blinker==1.8.2
build==1.2.2.post1
cachelib==0.13.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
docutils==0.21.2
executing==2.0.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Session==0.8.0
fonttools==4.54.1
gunicorn==22.0.0
idna==3.7
importlib_metadata==8.0.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.1
Jinja2==3.1.4
jsonpickle==3.2.2
jupyter_client==8.6.2
jupyter_core==5.7.2
keyring==25.2.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
modelpy-abm==0.0.3
more-itertools==10.3.0
msgspec==0.18.6
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nh3==0.2.18
numpy==2.0.0
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
pkginfo==1.10.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.2.0
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.0.3
readme_renderer==44.0
redis==5.0.7
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
six==1.16.0
stack-data==0.6.3
tokenize-rt==5.2.0
tornado==6.4.1
traitlets==5.14.3
twine==5.1.1
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
zipp==3.19.2
